Refactoring activation struct to activations module
