+++ delta/src/neuralnet/layers/conv_2d.rs
@@ -27,10 +27,11 @@
 //! OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 //! OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+use crate::activations::Activation;
 use crate::common::layer::Layer;
 use crate::common::shape::Shape;
 use crate::common::tensor_ops::Tensor;
-use crate::common::{Activation, LayerError};
+use crate::common::LayerError;
 
 /// A struct representing a 2D convolutional layer.
 #[derive(Debug)]
