+++ delta/src/neuralnet/layers/dense.rs
@@ -27,9 +27,10 @@
 //! OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 //! OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use crate::common::{Activation, Layer, LayerError, Optimizer, Shape, Tensor};
+use crate::common::{Layer, LayerError, Optimizer, Shape, Tensor};
 use log::debug;
 use serde_json;
+use crate::activations::Activation;
 
 /// A dense (fully connected) layer.
 #[derive(Debug)]
